Restructuring part 2
Divided account.py and the subparser in application.py into multiple files,
according to their subcommand. Every file has it's own subparser now.

The new files are:

- matrixctl/adduser.py
- matrixctl/adduser_jitsi.py
- matrixctl/deluser.py
- matrixctl/deluser_jitsi.py
- matrixctl/user.py
- matrixctl/users.py
- matrixctl/password_helpers.py (no subparser, just helper)

Deleted unused file:

- matrixctl/account.py
